fix : 型定義をroute.tsに移動
