+++ src/app/api/posts/[id]/route.ts
@@ -1,4 +1,5 @@
 import { db } from "@/lib/db";
+import { Post } from "@/types";
 import { NextRequest, NextResponse } from "next/server";
 
 export async function GET(
@@ -7,7 +8,7 @@ export async function GET(
 ) {
   try {
     const { id } = await params;
-    const post = await db.post.findUnique({
+    const post: Post | null = await db.post.findUnique({
       where: { id: parseInt(id) },
       include: {
         PostCategory: {
